+++ portal/front/portal-solution-desktop/src/main/ipc.ts
@@ -1,51 +1,63 @@
-import { join } from 'path'
-import { app, BrowserWindow, ipcMain, session } from 'electron'
+import { ipcMain, session, shell } from 'electron'
 import compressing from 'compressing'
 
 import { callEditor } from './utils'
+import { DownloadFileResult } from '../preload/type'
 
-const unzip = (fileName: string, filePath: string, suffix: string): Promise<void> => {
-  if (suffix === '.zip') {
-    const source = `${filePath}/${fileName}${suffix}`
-    console.log('unzip dest', source)
-    return compressing.zip.uncompress(source, filePath)
-  }
-  return Promise.reject(new TypeError(`不支持的文件类型：${suffix}`))
-}
+const setUpDownload = (): void => {
+  const downloadMap = new Map<
+    string,
+    {
+      resolve: (value: DownloadFileResult) => void
+      reject: (reason?: unknown) => void
+    }
+  >()
+  ipcMain.handle('downloadFile', async (_, url) => {
+    session.defaultSession.downloadURL(url)
+    return new Promise((resolve, reject) => {
+      downloadMap.set(url, {
+        resolve,
+        reject
+      })
+    })
+  })
 
-const registerWillDownload = (mainWindow: BrowserWindow): void => {
   session.defaultSession.on('will-download', (_, item) => {
     const fileName = item.getFilename()
-    const filePath = join(app.getPath('downloads'), fileName)
-    item.setSavePath(filePath)
 
-    mainWindow.webContents.send('downloadStatus', {
-      code: 0,
-      message: `${fileName}下载中`
-    })
+    item.once('done', (_, result) => {
+      if (result === 'completed') {
+        downloadMap.get(item.getURL())?.resolve({
+          fileName,
+          filePath: item.getSavePath()
+        })
+        return
+      }
 
-    item.once('done', () => {
-      mainWindow.webContents.send('downloadStatus', {
-        code: 1,
-        message: `${fileName}下载成功`
-      })
+      if (result === 'cancelled' || result === 'interrupted') {
+        downloadMap.get(item.getURL())?.reject(new Error(result))
+        return
+      }
     })
   })
 }
 
-export const initIPC = (mainWindow: BrowserWindow): void => {
-  registerWillDownload(mainWindow)
+export const initIPC = (): void => {
+  setUpDownload()
 
   ipcMain.handle('openEditor', (_, info) => {
     const { filePath, platform } = info
     return callEditor(platform, filePath)
   })
 
-  ipcMain.handle('getDownloadsPath', async () => {
-    return app.getPath('downloads')
+  ipcMain.handle('unzip', async (_, fileName, filePath) => {
+    if (fileName.endsWith('.zip')) {
+      return compressing.zip.uncompress(filePath, `${filePath.replace('/' + fileName, '')}`)
+    }
+    return Promise.reject(new TypeError(`不支持的文件类型：${filePath}`))
   })
 
-  ipcMain.handle('unzip', async (_, fileName, filePath, suffix) => {
-    return unzip(fileName, filePath, suffix)
+  ipcMain.handle('openFile', async (_, filePath) => {
+    return shell.openPath(filePath)
   })
 }