+++ portal/front/portal-solution/src/components/lowcode/ProjectList/index.tsx
@@ -16,9 +16,10 @@ import moment from 'moment'
 import { useEffect, useState } from 'react'
 import { Modal } from 'react-icecream'
 
+import { Platform } from 'utils/electron'
 import { useData } from 'components/lowcode/ProjectList/useData'
 
-import { Platform, ProjectInfo } from './type'
+import { ProjectInfo } from './type'
 
 import './style.less'
 
@@ -40,26 +41,20 @@ const platformOptions = [
 const sceneMap = {
   1: '视频营销/统一消息推送'
 }
+const platformMap = {
+  ios: 'iOS',
+  android: 'Android'
+}
 
 export function LowcodeProjectList() {
   const [searchSceneType, setSearchSceneType] = useState<number>(1)
   const [searchPlatform, setSearchPlatform] = useState<Platform | 'all'>('all')
-  const [downloadsPath, setDownloadsPath] = useState('')
 
   const [originalRecords, setOriginalRecords] = useState<ProjectInfo[]>([])
   const [filteredOriginalRecords, setFilteredOriginalRecords] = useState<ProjectInfo[]>([])
   const { records, loading, currentPage, pageSize, total, setPageInfo, setLoading } = useData(filteredOriginalRecords)
 
   const [searchProjectName, setSearchProjectName] = useState('')
-
-  /**
-   * 从 electron main 进程获取下载路径
-   */
-  useEffect(() => {
-    window.electronBridgeApi.getDownloadsPath().then(value => {
-      setDownloadsPath(value)
-    })
-  }, [])
 
   /**
    * 从缓存中读取项目列表
@@ -83,17 +78,18 @@ export function LowcodeProjectList() {
    * @param record
    */
   const onOpenEditor = (type: Platform, record: ProjectInfo) => {
+    const packageInfo = record.package[type.toLowerCase()]
+    console.log('packageInfo', packageInfo)
     window.electronBridgeApi.unzip(
-      record.name,
-      downloadsPath,
-      '.zip'
+      packageInfo?.fileName,
+      packageInfo?.filePath
     ).then(() => {
       Modal.success({
         content: '解压成功'
       })
-      return window.electronBridgeApi.openEditor({
+      return window.electronBridgeApi?.openEditor({
         platform: type,
-        filePath: `${downloadsPath}/${record.name}`
+        filePath: `${packageInfo?.filePath.replace('.zip', '')}`
       })
     }).then(() => {
       Modal.success({
@@ -110,15 +106,22 @@ export function LowcodeProjectList() {
   /**
    * 搜索
    */
-  const onSearch = async () => {
+  const onSearch = async (info: {
+    sceneType?: number
+    platform?: Platform | 'all'
+    projectName?: string
+  }) => {
+    const {
+      sceneType, projectName, platform
+    } = info
     setLoading(true)
     const list = originalRecords
-      .filter(item => item.sceneType === searchSceneType)
+      .filter(item => item.sceneType === sceneType)
       .filter(item => {
-        if (!searchPlatform || searchPlatform === 'all') return true
-        return item.platform.includes(searchPlatform)
+        if (!platform || platform === 'all') return true
+        return !!item.package[platform.toLowerCase()]
       })
-      .filter(item => item.name.includes(searchProjectName))
+      .filter(item => item.name.includes(projectName || ''))
     await new Promise(resolve => setTimeout(resolve, 300))
     setLoading(false)
     setFilteredOriginalRecords(list)
@@ -135,7 +138,11 @@ export function LowcodeProjectList() {
             value={searchSceneType}
             onChange={(value: number) => {
               setSearchSceneType(value)
-              onSearch()
+              onSearch({
+                projectName: searchProjectName,
+                platform: searchPlatform,
+                sceneType: value
+              })
             }}
           >
             {sceneTypeOptions}
@@ -146,7 +153,11 @@ export function LowcodeProjectList() {
             value={searchPlatform}
             onChange={(value: Platform | 'all') => {
               setSearchPlatform(value)
-              onSearch()
+              onSearch({
+                projectName: searchProjectName,
+                platform: value,
+                sceneType: searchSceneType
+              })
             }}
           >
             {platformOptions}
@@ -161,7 +172,11 @@ export function LowcodeProjectList() {
             inputProps={{
               onKeyPress: event => {
                 if (event.key.toLowerCase() === 'enter') {
-                  onSearch()
+                  onSearch({
+                    projectName: searchProjectName,
+                    platform: searchPlatform,
+                    sceneType: searchSceneType
+                  })
                 }
               }
             }}
@@ -182,7 +197,32 @@ export function LowcodeProjectList() {
           },
           {
             title: '链路',
-            render: (_value, record) => <Link style={{ wordBreak: 'break-all' }}>{downloadsPath}/{record.name}</Link>
+            render: (_value, record: ProjectInfo) => {
+              const { android, ios } = record.package
+              const androidLink = android?.filePath?.replace(`/${android?.fileName}`, '')
+              const iosLink = ios?.filePath?.replace(`/${ios?.fileName}`, '')
+
+              return (
+                <div>
+                  {
+                    android && <div>
+                      Android: <Link
+                        style={{ wordBreak: 'break-all' }}
+                        onClick={() => window.electronBridgeApi.openFile(androidLink || '')}
+                      >{androidLink}</Link>
+                    </div>
+                  }
+                  {
+                    ios && <div>
+                      iOS: <Link
+                        style={{ wordBreak: 'break-all' }}
+                        onClick={() => window.electronBridgeApi.openFile(iosLink || '')}
+                      >{iosLink}</Link>
+                    </div>
+                  }
+                </div>
+              )
+            }
           },
           {
             title: '场景分类',
@@ -200,17 +240,26 @@ export function LowcodeProjectList() {
           },
           {
             title: '端类型',
-            accessor: 'platform',
-            render: value => value.join('、')
+            accessor: 'package',
+            render: value => {
+              const result: string[] = []
+              if (value.android) {
+                result.push(platformMap.android)
+              }
+              if (value.ios) {
+                result.push(platformMap.ios)
+              }
+              return result.join('、')
+            }
           },
           {
             title: '操作',
-            render: (_, record) => <Dropdown
+            render: (_, record: ProjectInfo) => <Dropdown
               trigger="click"
               overlay={
                 <Menu>
-                  {record.platform?.includes('Android') && <MenuItem onClick={() => onOpenEditor('Android', record)}>Android</MenuItem>}
-                  {record.platform?.includes('iOS') && <MenuItem onClick={() => onOpenEditor('iOS', record)}>iOS</MenuItem>}
+                  {record.package.android && <MenuItem onClick={() => onOpenEditor('android', record)}>Android</MenuItem>}
+                  {record.package.ios && <MenuItem onClick={() => onOpenEditor('ios', record)}>iOS</MenuItem>}
                 </Menu>
               }
             >