+++ portal/front/portal-solution/src/components/lowcode/ProjectList/type.ts
@@ -1,4 +1,9 @@
-export type Platform = Parameters<typeof window.electronBridgeApi.openEditor>[0]['platform']
+import { Platform } from 'utils/electron'
+
+interface PackageInfo {
+  fileName: string
+  filePath: string
+}
 
 export type ProjectInfo = {
   /**
@@ -23,8 +28,7 @@ export type ProjectInfo = {
    */
   createTime: number
   /**
-   * 端类型
-   * 'Android' | 'iOS'
+   * 包信息
    */
-  platform: Platform[]
+  package: Partial<Record<Platform, PackageInfo>>
 }