+++ portal/front/portal-solution/src/constants/is.ts
@@ -1,1 +1,1 @@
-export const isElectron = !!window.electronBridgeApi
+export const isElectron = window.navigator.userAgent.toLowerCase().indexOf('electron') > -1